+++ common/src/main/java/net/conczin/immersive_furniture/client/gui/ArtisansWorkstationLibraryScreen.java
@@ -643,6 +643,7 @@ public class ArtisansWorkstationLibraryScreen extends ArtisansWorkstationScreen 
             if (data.hasSounds()) tags.add("has_sounds");
             if (data.canSit()) tags.add("can_sit");
             if (data.canSleep()) tags.add("can_sleep");
+            if (data.hasDisplayItems()) tags.add("has_display_items");
             if (data.getUniqueSolidStates().size() > 1) tags.add("has_states");
             if (!data.dependencies.isEmpty()) tags.add("has_dependencies");
             if (!data.sources.isEmpty()) tags.add("has_modded_textures");